pyproject.toml
[tool.poetry]
name = "iwell"
version = "0.1.0"
description = ""
authors = ["Brock Friedrich <brock@driftwoodenergy.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "*"
requests = "*"
requests-oauthlib = "==1.0.0"
pandas = "*"
sqlalchemy = "==1.2.18"
sentry-sdk = "*"
oauthlib = "*"
click = "*"
psycopg2-binary = "^2.8"
flask = "^1.1"
Flask-SQLAlchemy = "^2.4"
flask-debugtoolbar = "*"
flask-migrate = "*"
Flask-RESTful = "*"
python-dotenv = "^0.10.3"
attrdict = "^2.0"
celery = "^4.3"
redis = "^3.3"
flower = "^0.9.3"
requests_mock = "^1.7"
tomlkit = "^0.5.8"
logutils = "^0.3.5"


[tool.poetry.dev-dependencies]
black = { version = "*", allows-prereleases = true }
mypy = "*"
pylint = "*"
ipython = "*"
matplotlib = "*"
pytest = "^5.2"
requests-mock = "^1.7"
pytest-cov = "^2.8"
coverage = "^4.5"


[tool.poetry.scripts]
iwell = "iwell.manage:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


